add pointer to struct conversion
