+++ test/fixtures/scope_with_pointer_param.yml
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
-# Copyright 2025 Joel E. Anderson
+# Copyright 2025-2026 Joel E. Anderson
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,15 +19,17 @@ classes:
     namespace: "wrapture_test"
     wrapped:
       c:
-        name: "bullet_struct"
-        includes: "bullet_struct.h"
+        pointer:
+          name: "bullet_struct"
+          includes: "bullet_struct.h"
     type: "pointer"
   - name: "Rifle"
     namespace: "wrapture_test"
     wrapped:
       c:
-        name: "rifle_struct"
-        includes: "rifle_struct.h"
+        pointer:
+          name: "rifle_struct"
+          includes: "rifle_struct.h"
     functions:
       - name: "load"
         params:
